Initial commit
This is the version of my Hearthstone card tracker from many months ago.
I've decided to improve it/make a new one (still deciding at time of
writing)

I include it for posterity, and to perhaps refer back to.

Also, it would be nice to see how different my Python looks now to how
it did.