tracker V1.0!
- decks can be reset by passing a "remember" flag to `Deck.take_card`
when we want to be able to put that card back in the deck later (see
`Deck.reset`)
- fixed abnormal draws bugging, for example thoughtsteal, Chromaggus
etc.
- fixed getting confused by The Coin (I was testing in solo adventures
mostly, where you always go first, so this got missed for a while)
- added `Deck.clear` (self-explanatory) and `Deck.reset`, which will add
in the cards remembered while being taken out (see `Deck.take_card`)
- commented on what the regexi actually pertain to, and modified one of
them to recognise when we're initially drawing compared to when
Chromaggus is being fruity
- ctrl+R will now clear the deck being edited with deck_builder

Overall, I feel that the basic functionality is there for both the
tracker and the builder, so I think it's ready to be called V1.0!
Further updates will be described as "features" and likely done in their
own branch, so it'll be a bit easier to keep track of what's changing
from here on, rather than everything being bundled into the master
branch.